hdl: Updated testbench to use part models.
Added a VCTCXO model and a 16-bit DAC model for the bladeRF.
The mm_driver.vhd has a more reasonable frequency adjust state.
Simulations will need to be tweaked because they take way too
long. Perhaps best to not use the device models because the changes
are too small to see over a reasonable simulation time.