nios: Added PPS calibration helper functions and modified ppscal_isr()
Added ppscal_reset_counters(bool), which allows application code to reset
the PPS calibration counters without having to know the register address
to do so.

Changed the ppscal_enable_isr(bool) function from being statically defined
and declared in devices.c to being declared in devices.h for use by application
code.

Modified the ppscal_isr() function to disable future interrupts from the PPS
calibration module, clear existing interrupts, and reset the PPS counters.
It is the job of the application to re-enable interrupts and to take the
counters out of reset.